Refactor USBTMC/GPIB
