+++ interface/gpib.py
@@ -1,20 +1,12 @@
-import serial, time
+from interface.usbtmc import USBTMC
 
-class GPIB:
-    # Current serial connection
-    _ser = None
-
+class GPIB(USBTMC):
     def __init__(self,
             serial_port: str,
             gpib_addr: int,
             baud: int = 19200
         ):
-        # Open the serial connection
-        print(f'Opening serial connection on {serial_port}...', end=' ', flush=True)
-        self._ser = serial.Serial(serial_port, baud, timeout=2)
-        self._ser.reset_input_buffer()
-        self._ser.reset_output_buffer()
-        print('Done.')
+        super().__init__(serial_port, baud)
 
         # Set address and timeout
         print('Bridge Version:', self.query('++ver'))
@@ -25,23 +17,5 @@ class GPIB:
         self.send_command('++read')
         self._ser.read()
 
-        # Query model
-        model = self.query('*idn?')
-        print('Device Model:', model)
-
-    def stop(self):
-        """Closes the serial connection."""
-        self._ser.close()
-
-
-    ##### Utility Functions #####
-    def send_command(self, command: str) -> None:
-        """Send a command to the GPIB device."""
-        self._ser.write((command + '\n').encode('utf-8'))
-        time.sleep(1.0)
-
-    def query(self, command: str) -> str:
-        """Send a command to the GPIB device, and return its response."""
-        self.send_command(command)
-        response = self._ser.readline().strip()
-        return response.decode('utf-8')
+        # Display device model
+        print('Device Model:', self.get_name())
