+++ interface/oscilloscope.py
@@ -1,0 +1,5 @@
+from interface.usbtmc import USBTMC 
+
+class Oscilloscope(USBTMC):
+    """USBMTC interface for 
+    pass
