--- interface/rigol.py
@@ -1,4 +1,0 @@
-from interface.gpib import GPIB
-
-class Rigol(GPIB):
-    pass
